+++ src/webserver/src/WebSocket.h
@@ -51,7 +51,7 @@ class CWebSocket : public CLibSocket {
 
 		void SendContent(const char* szStdResponse, const void* pContent, uint32 dwContentSize);
 		void SendData(const void* pData, uint32 dwDataSize);
-		void SendHttpHeaders(const char * szType, bool use_gzip, uint32 content_len, int session_id);
+		void SendHttpHeaders(const char * szType, bool use_gzip, uint32 content_len, uint64_t session_id);
 
 		CWebServerBase *m_pParent;
